move pantry FEC testing comps
