+++ Saucier720-app/src/app/testing/pantry/FEC/new-pantry-item-button/new-pantry-item-button.component.spec.cy.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { NewPantryItemButtonComponent } from './new-pantry-item-button.component';
+import { NewPantryItemButtonComponent } from '../../../../pantry/FEC/new-pantry-item-button/new-pantry-item-button.component';
 
 describe('NewPantryItemButtonComponent', () => {
   let component: NewPantryItemButtonComponent;
@@ -15,9 +15,5 @@ describe('NewPantryItemButtonComponent', () => {
     fixture = TestBed.createComponent(NewPantryItemButtonComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
   });
 });
